Auto create self-signed keys
If the USE_SSl var is true and the private key and certificate does not
exist, then automatically create self signed keys. The common name would
be invalid so you would have to change it later. This addition was added
so the server would not crash if USE_SSL is true and keys do not exist.